move func tests
